requirements.txt
streamlit
psycopg2-binary
python-dotenv
PyPDF2
google-generativeai
langgraph
langchain-core
dotenv